pyproject.toml
[project]
name = "dope"
version = "0.1.0"
description = ""
authors = [{ name = "Martin Gran", email = "martgra@gmail.com" }]
requires-python = ">=3.13"
readme = "README.md"
dependencies = [
    "anytree>=2.13.0",
    "gitpython>=3.1.44",
    "platformdirs>=4.3.7",
    "pydantic>=2.11.3",
    "pydantic-ai>=0.1.2",
    "pydantic-settings>=2.9.0",
    "pyyaml>=6.0.2",
    "questionary>=2.1.0",
    "rich>=14.0.0",
    "typer>=0.15.3",
]

[project.urls]
Documentation = "https://github.com/martgra/dope/README.md"
Source = "https://github.com/martgra/dope"


[dependency-groups]
dev = [
    "deptry>=0.23.1",
    "detect-secrets>=1.5.0",
    "prek>=0.2.2",
    "pylint>=3.3.8",
    "ruff>=0.13.2",
    "ty>=0.0.1a29",
    "vulture>=2.14",
]
test = [
    "faker>=37.1.0",
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
]

[tool.uv]
default-groups = [
    "dev",
    "test",
]

[project.scripts]
dope = "dope.cli.main:run_cli"

[tool.hatch.build.targets.sdist]
include = ["dope"]

[tool.hatch.build.targets.wheel]
include = ["dope"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
exclude = ["tests/*", "vulture_whitelist.py"]

[tool.ruff.lint]
extend-ignore = ["E203", "D107", "D100"]
select = ["E", "W", "D", "F", "UP", "B", "SIM","I"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["D104"]
"app/core/settings.py" = ["D101", "D102", "D103", "D104", "D105", "D106", "D107"]
"**/agents/*.py" = ["D417"]
"**/**/prompts.py" = ["E501"]
"**/cli/*.py" = ["D103"]
"app/models/internal.py" = ["D101", "D102", "D103", "D105"]

[tool.ruff.lint.isort]
known-first-party = ["dope"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.pylint]
max-line-length = 100
fail-under = 9.5
disable = ["E0401", "C0114", "R0903", "E0237","E1142","E0014","E1300","E0013","E1310","E1307","E2502","E6005","E6004","E0116","E0108","E0241","E1303","E0102","E0100","E0605","E0604","E0304","E0308","E2510","E2513","E2514","E2512","E2515","E0309","E0305","E0303","E1206","E1205","E0704","E1304","E1302","E4703","E0213","E0107","E0115","E0117","E0103","E0711","E0643","E0402","E1132","E0106","E0101","E0104","E1519","E1520","E0001","E1306","E1305","E0112","E1301","E0603","E0602","E0302","E0118","E1700","E0105","W1401","W0129","W0199","W3201","W1302","W1300","W1501","W0211","W0702","W0711","W1502","W0718","W0719","W0640","W0160","W0102","W0705","W0109","W1308","W0130","W1641","W0123","W0122","W0106","W1309","W0511","W1515","W1305","W1310","W0604","W0603","W0602","W1404","W0406","W1405","W1508","W1113","W1202","W1203","W1201","W0150","W1518","W0410","W1303","W0131","W0177","W3301","W2402","W0133","W0104","W0212","W0707","W0622","W2901","W1406","W0404","W0127","W1509","W1510","W0245","W0706","W0012","W0108","W0107","W0301","W1514","W0613","W1304","W1301","W0611","W0612","W0120","W2101","W2601","W0401","C0202","C0198","C1901","C0201","C0501","C0206","C0199","C0112","C0415","C2701","C0103","C0301","C2201","C0115","C0304","C0116","C0114","C0410","C0321","C2403","C2401","C0205","C0121","C0303","C0131","C0105","C0132","C0412","C0123","C3002","C2801","C3001","C0113","C0208","C0414","C0411","C0413","R0133","R0124","R6003","R1701","R6002","R6104","R1717","R1728","R1715","R1714","R1730","R1731","R1718","R1722","R1706","R1732","R5501","R2044","R1710","R0123","R2004","R0202","R1723","R1724","R1720","R1705","R6301","R0203","R0206","R1704","R1719","R1703","R1725","R1260","R0913","R0916","R0912","R0914","R1702","R0904","R0911","R0915","R1707","R1721","R1733","R1736","R1729","R1735","R1734","R6201","R0205","R0022","R1711"]

[tool.deptry]
known_first_party = ["dope"]

[tool.pytest]
testpaths = ["tests"]

[tool.coverage.run]
branch = true
source = ["dope"]

[tool.vulture]
make_whitelist = true
min_confidence = 80
paths = ["dope"]
sort_by_size = true
